+++ src/app/budget/components/line-menu/line-menu.component.spec.ts
@@ -2,24 +2,24 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { LineMenuComponent } from './line-menu.component';
 
-describe('LineMenuComponent', () => {
-  let component: LineMenuComponent;
-  let fixture: ComponentFixture<LineMenuComponent>;
+// describe('LineMenuComponent', () => {
+//   let component: LineMenuComponent;
+//   let fixture: ComponentFixture<LineMenuComponent>;
 
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ LineMenuComponent ]
-    })
-    .compileComponents();
-  }));
+//   beforeEach(async(() => {
+//     TestBed.configureTestingModule({
+//       declarations: [ LineMenuComponent ]
+//     })
+//     .compileComponents();
+//   }));
 
-  beforeEach(() => {
-    fixture = TestBed.createComponent(LineMenuComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
+//   beforeEach(() => {
+//     fixture = TestBed.createComponent(LineMenuComponent);
+//     component = fixture.componentInstance;
+//     fixture.detectChanges();
+//   });
 
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
+//   it('should create', () => {
+//     expect(component).toBeTruthy();
+//   });
+// });