+++ src/app/core/components/undo-snackbar/undo-snackbar.component.spec.ts
@@ -2,24 +2,24 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { UndoSnackbarComponent } from './undo-snackbar.component';
 
-describe('UndoSnackbarComponent', () => {
-  let component: UndoSnackbarComponent;
-  let fixture: ComponentFixture<UndoSnackbarComponent>;
+// describe('UndoSnackbarComponent', () => {
+//   let component: UndoSnackbarComponent;
+//   let fixture: ComponentFixture<UndoSnackbarComponent>;
 
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ UndoSnackbarComponent ]
-    })
-    .compileComponents();
-  }));
+//   beforeEach(async(() => {
+//     TestBed.configureTestingModule({
+//       declarations: [ UndoSnackbarComponent ]
+//     })
+//     .compileComponents();
+//   }));
 
-  beforeEach(() => {
-    fixture = TestBed.createComponent(UndoSnackbarComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
+//   beforeEach(() => {
+//     fixture = TestBed.createComponent(UndoSnackbarComponent);
+//     component = fixture.componentInstance;
+//     fixture.detectChanges();
+//   });
 
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
+//   it('should create', () => {
+//     expect(component).toBeTruthy();
+//   });
+// });